Transition from SVN to Git
Added implicit conversion to base types to all primary types (UInt32 et
al)
Added implicit conversion from System::String to const char*

+++ src/libXFX/Viewport.cpp
@@ -96,7 +96,7 @@ namespace XFX
 
 		int Viewport::GetHashCode() const
 		{
-			return ((int)getAspectRatio() + Height + MaxDepth + MinDepth + Width + X + Y);
+			return ((int)getAspectRatio() + Height + (int)MaxDepth + (int)MinDepth + Width + X + Y);
 		}
 
 		int Viewport::GetType() const
